Fixed formatting and lint. Fixed a warning flagged by lint.
