+++ src/styles/_fonts.scss
@@ -3,7 +3,7 @@ $font-sans:
   system-ui,
   -apple-system,
   sans-serif;
-  
+
 $font-display:
   "Segoe UI",
   system-ui,
