+++ src/styles/_theme.scss
@@ -23,7 +23,7 @@
   --color-border-hover: #ffffff;
 }
 
-:root[data-theme='dark'] {
+:root[data-theme="dark"] {
   // Brand accents — lifted for contrast on dark surfaces
   --color-primary: #4aa3d4;
   --color-secondary: #6bb8de;
