+++ src/styles/components/_carousel.scss
@@ -1,4 +1,4 @@
-@use '../variables' as *;
+@use "../variables" as *;
 
 .carousel {
   position: relative;
