+++ src/styles/components/_contact-form.scss
@@ -1,4 +1,4 @@
-@use '../variables' as *;
+@use "../variables" as *;
 
 // Contact form
 .contact {
