+++ src/styles/components/_footer.scss
@@ -1,4 +1,4 @@
-@use '../variables' as *;
+@use "../variables" as *;
 
 .footer {
   background: var(--color-primary);
