+++ src/styles/components/_project-card.scss
@@ -1,4 +1,4 @@
-@use '../variables' as *;
+@use "../variables" as *;
 
 .project-card {
   display: block;
