+++ src/styles/components/_theme-menu.scss
@@ -1,6 +1,6 @@
 // Theme preference dropdown in the site navbar.
 
-@use '../variables' as *;
+@use "../variables" as *;
 
 .theme-menu {
   position: relative;
