+++ src/styles/main.scss
@@ -1,13 +1,13 @@
 // Entry stylesheet — re-exports style partials only.
 
-@use 'normalize' as *;
-@use 'reset' as *;
-@use 'theme' as *;
-@use 'base' as *;
-@use 'layout' as *;
-@use 'components/carousel' as *;
-@use 'components/contact-form' as *;
-@use 'components/footer' as *;
-@use 'components/nav' as *;
-@use 'components/project-card' as *;
-@use 'components/theme-menu' as *;
+@use "normalize" as *;
+@use "reset" as *;
+@use "theme" as *;
+@use "base" as *;
+@use "layout" as *;
+@use "components/carousel" as *;
+@use "components/contact-form" as *;
+@use "components/footer" as *;
+@use "components/nav" as *;
+@use "components/project-card" as *;
+@use "components/theme-menu" as *;
